// src/presentation/components/Logos/interface/iLogos.d.ts
export type LogosType =
  | "next"
  | "webpack"
  | "vue"
  | "xd"
  | "wordpress"
  | "typescript"
  | "sass"
  | "indesign"
  | "photoshop"
  | "marvel"
  | "html"
  | "react"
  | "prettier"
  | "css"
  | "figma"
  | "eslint"
  | "git"
  | "affinity"
  | "illustrator";

export interface ILogos {
  logos: LogosType[];
}